Trekker ut tester som ikke er spesifikt for næringsdrivende til egen fil
